Extract shared test utilities to test_utils module
Consolidate duplicated timeout configurations and the wait_for_console_log helper function into a new tests/test_utils.py module. This eliminates code duplication across 9 test files while maintaining identical functionality. All test files now import these shared utilities from the centralized module.
