+++ tests/audio_flow_test.py
@@ -33,8 +33,7 @@ from typing import Any, Callable
 import pytest
 from playwright.sync_api import Page, expect
 
-DEFAULT_TIMEOUT = int(os.getenv("TEST_TIMEOUT", "30000"))
-TEST_TIMEOUT = int(os.getenv("TEST_TIMEOUT", "5000"))
+from tests.test_utils import DEFAULT_TIMEOUT, TEST_TIMEOUT, wait_for_console_log
 
 
 @pytest.mark.record_har
@@ -57,20 +56,6 @@ def test_audio_flow(page: Page) -> None:
         logs.append({"type": msg.type, "text": msg.text})
 
     page.on("console", on_console)
-
-    def wait_for_console_log(
-        predicate: Callable[[str], bool], start_idx: int, timeout_ms: int = TEST_TIMEOUT
-    ) -> None:
-        """Helper to poll until a matching console log arrives or timeout expires."""
-        start_time = time.time()
-        while (time.time() - start_time) * 1000 < timeout_ms:
-            if any(predicate(log["text"].lower()) for log in logs[start_idx:]):
-                return
-            page.wait_for_timeout(50)  # Micro-poll for event loop progression
-        pytest.fail(
-            "Timed out waiting for expected console log matching "
-            f"predicate after {timeout_ms}ms"
-        )
 
     try:
         # Start CDP session for V8 JS coverage
