+++ tests/back_flow_test.py
@@ -36,8 +36,7 @@ import pytest
 from playwright.sync_api import Page, expect
 
 # Configuration for stability in different environments
-DEFAULT_TIMEOUT = int(os.getenv("TEST_TIMEOUT", "30000"))
-TEST_TIMEOUT = int(os.getenv("TEST_TIMEOUT", "10000"))
+from tests.test_utils import DEFAULT_TIMEOUT, TEST_TIMEOUT, wait_for_console_log
 
 
 def test_back_flow(page: Page) -> None:
@@ -64,22 +63,6 @@ def test_back_flow(page: Page) -> None:
         logs.append({"type": msg.type, "text": msg.text})
 
     page.on("console", on_console)
-
-    def wait_for_console_log(
-        predicate: Callable[[str], bool],
-        start_idx: int,
-        timeout_ms: int = TEST_TIMEOUT,
-    ) -> None:
-        """Helper to poll until a matching console log arrives or timeout expires."""
-        start_time = time.time()
-        while (time.time() - start_time) * 1000 < timeout_ms:
-            if any(predicate(log["text"].lower()) for log in logs[start_idx:]):
-                return
-            page.wait_for_timeout(50)  # Micro-poll for event loop progression
-        pytest.fail(
-            "Timed out waiting for expected console log matching "
-            f"predicate after {timeout_ms}ms"
-        )
 
     try:
         # Start CDP session for V8 JS coverage
