+++ tests/test_utils.py
@@ -1,0 +1,36 @@
+# Copyright (C) 2025 Egor Kostan
+# SPDX-License-Identifier: GPL-3.0-or-later
+# tests/test_utils.py
+"""
+Shared utility functions and helpers for SkyLockAssault Playwright E2E tests.
+"""
+
+import os
+import time
+from typing import Callable
+
+import pytest
+from playwright.sync_api import Page
+
+# Shared timeout configurations across test suites
+DEFAULT_TIMEOUT = int(os.getenv("TEST_TIMEOUT", "30000"))
+TEST_TIMEOUT = int(os.getenv("TEST_TIMEOUT", "5000"))
+
+
+def wait_for_console_log(
+    logs: list[dict[str, str]],
+    predicate: Callable[[str], bool],
+    start_idx: int,
+    page: Page,
+    timeout_ms: int = TEST_TIMEOUT,
+) -> None:
+    """Helper to poll until a matching console log arrives or timeout expires."""
+    start_time = time.time()
+    while (time.time() - start_time) * 1000 < timeout_ms:
+        if any(predicate(log["text"].lower()) for log in logs[start_idx:]):
+            return
+        page.wait_for_timeout(50)  # Micro-poll for event loop progression
+    pytest.fail(
+        "Timed out waiting for expected console log matching "
+        f"predicate after {timeout_ms}ms"
+    )
